Move cluster and storage selections from Resources step to General, rename Resources to Namespaces
